Java Experiments with various OOPS concept
Encapsulation: Bundling data and methods that operate on the data within a single unit or class, and restricting access to some of the object's components.

Abstraction: Hiding complex implementation details and exposing only the necessary parts of the object through well-defined interfaces.

Inheritance: Enabling new classes to inherit properties and behaviors from existing classes, promoting code reusability and hierarchical relationships.

Polymorphism: Allowing objects to be treated as instances of their parent class, and enabling methods to behave differently based on the object's actual class (method overriding and overloading).

Classes and Objects: A class defines a blueprint for objects. Objects are instances of classes, representing real-world entities or concepts with attributes (data) and behaviors (methods).